por implementar metodo de pago

+++ src/app/modules/recepcionista-role/pagos/pagos-pages/pagos-pages.component.ts
@@ -12,9 +12,12 @@ import { InputNumber } from 'primeng/inputnumber';
 import { AutoComplete } from 'primeng/autocomplete';
 import { MessagesService } from '@core/services/message/messages.service';
 import { DialogModule } from 'primeng/dialog';
+import { CommonModule } from '@angular/common';
+
 @Component({
   selector: 'app-pagos-pages',
   imports: [
+    CommonModule,
     InputGroupAddonModule,
     InputGroupModule,
     InputTextModule,
@@ -25,7 +28,7 @@ import { DialogModule } from 'primeng/dialog';
     DividerModule,
     InputNumber,
     AutoComplete,
-    DialogModule
+    DialogModule,
   ],
   templateUrl: './pagos-pages.component.html',
   styleUrls: ['./pagos-pages.component.css'],
@@ -33,7 +36,7 @@ import { DialogModule } from 'primeng/dialog';
 export class PagosPagesComponent implements OnInit {
   consulta: number | undefined;
   id: number | undefined;
-  pagoId: number | undefined;
+  pagoId: number[] = [];
   monto: number | undefined;
   items: any[] = [];
   products: any[] = [];
@@ -41,6 +44,7 @@ export class PagosPagesComponent implements OnInit {
   visiblePagar: boolean = false;
   dni: string = '';
   dniError: boolean = false;
+  seleccionPrevia: boolean = false; // Nueva variable de estado
 
   constructor(
     private pagosService: PagosService,
@@ -67,19 +71,24 @@ export class PagosPagesComponent implements OnInit {
   }
 
   onSelectOrden(event: any): void {
-    this.pagosService.deleteSeleccionar().subscribe({
-      next: (data: string)=>{
-        console.log(data);
-      }
-    })
+    if (this.seleccionPrevia && this.id !== undefined) {
+      this.pagosService.deleteSeleccionar().subscribe({
+        next: (data: string) => {
+          console.log(data);
+        }
+      });
+    }
     this.id = event.idOrden;
-    if (this.id !== undefined) {
-      
+    this.seleccionPrevia = true; // Actualiza la variable de estado
 
+    if (this.id !== undefined) {
       this.pagosService.postSeleccionar(this.id).subscribe({
         next: (data: string) => {
           console.log(data);
-          this.messagesService.successMessage('Orden seleccionada', 'La orden ha sido seleccionada con éxito');
+          this.messagesService.successMessage(
+            'Orden seleccionada',
+            'La orden ha sido seleccionada con éxito',
+          );
           this.pagosService.getListaCarrito().subscribe({
             next: (data: any) => {
               this.products = data;
@@ -103,40 +112,48 @@ export class PagosPagesComponent implements OnInit {
     if (this.consulta !== undefined) {
       this.pagosService.postAgregarConsulta(this.consulta).subscribe({
         next: (data: number) => {
-          this.pagoId = data;
-          this.messagesService.successMessage('Consulta añadida', 'La consulta ha sido añadida con éxito');
+          this.pagoId.push(data);
+          this.messagesService.successMessage(
+            'Consulta añadida',
+            'La consulta ha sido añadida con éxito',
+          );
           this.recargarCarrito();
         },
         error: (err) => {
           console.error('Error al añadir consulta', err);
-        }
+        },
       });
     }
   }
 
   eleminarConsulta(): void {
-    if (this.pagoId !== undefined) {
-      this.pagosService.deleteConsulta(this.pagoId).subscribe({
-        next: () => {
-          this.messagesService.successMessage('Consulta eliminada', 'La consulta ha sido eliminada con éxito');
-          this.recargarCarrito();
-        },
-        error: (err) => {
-          console.error('Error al eliminar consulta', err);
-        }
+    if (this.pagoId !== undefined && this.pagoId.length > 0) {
+      let eliminacionesCompletadas = 0;
+      this.pagoId.forEach((id) => {
+        this.pagosService.deleteConsulta(id).subscribe({
+          next: () => {
+            eliminacionesCompletadas++;
+            if (eliminacionesCompletadas === this.pagoId.length) {
+              this.messagesService.successMessage(
+                'Consulta eliminada',
+                'La consulta ha sido eliminada con éxito',
+              );
+              this.recargarCarrito();
+              
+            }
+            
+          },
+          error: (err) => {
+            console.error('Error al eliminar consulta', err);
+          },
+        });
       });
-    } else if (this.pagoId === undefined) {
-      this.pagosService.deleteConsulta(0).subscribe({
-        next: () => {
-          this.messagesService.successMessage('Consulta eliminada', 'La consulta ha sido eliminada con éxito');
-          this.recargarCarrito();
-        },
-        error: (err) => {
-          console.error('Error al eliminar consulta', err);
-        }
-      });
-    }else{
-      this.messagesService.errorMessage('Error', 'No se ha podido eliminar la consulta');
+      
+    }  else {
+      this.messagesService.errorMessage(
+        'Error',
+        'No se ha podido eliminar la consulta',
+      );
     }
   }
 
@@ -153,14 +170,87 @@ export class PagosPagesComponent implements OnInit {
   }
 
   calculateMontoTotal(): void {
-    this.monto = this.products.reduce((total, product) => total + product.subtotal, 0);
+    this.monto = this.products.reduce(
+      (total, product) => total + product.subtotal,
+      0,
+    );
   }
   validateDni(): void {
-    const dniPattern = /^[0-9]{8}$/
-    this.dniError = !dniPattern.test(this.dni); 
+    const dniPattern = /^[0-9]{8}$/;
+    this.dniError = !dniPattern.test(this.dni);
   }
 
-  confirmarPago(){
-    77
+  confirmarPago() {
+    this.pagosService.validarDNI(this.dni).subscribe({
+      next: (data: number) => {
+        this.pagosService.postConfirmar(data).subscribe({
+          next: (data: any) => {
+            this.messagesService.successMessage(
+              'Pago confirmado',
+              'El pago ha sido confirmado con éxito',
+            );
+            this.pagosService.getListaPagosPendientes().subscribe({
+              next: (data: any) => {
+                this.ordenes = data;
+                this.items = data.map((pago: any) => ({ idOrden: pago.idOrden }));
+              },
+              error: (err) => {
+                console.error('Error al cargar órdenes de pago', err);
+              },
+            });
+            this.recargarCarrito();
+            this.visiblePagar = false;
+          },
+          error: (err) => {
+            this.messagesService.errorMessage(
+              'Error',
+              'No se ha podido confirmar el pago',
+            );
+          },
+        });
+      },
+      error: (err) => {
+        this.messagesService.errorMessage(
+          'Error',
+          'No se ha podido validar el DNI',
+        );
+      },
+    });
+  }
+
+  expirarorden() {
+    if (this.id !== undefined) {
+      this.pagosService.expirarOrdenPago(this.id).subscribe({
+        next: (data: any) => {
+          this.messagesService.successMessage(
+            'Orden expirada',
+            'La orden ha sido cancelada con éxito',
+          );
+          this.pagosService.deleteSeleccionar().subscribe({
+            next: (data: string) => {
+              console.log(data);
+            },
+          });
+          this.pagosService.getListaPagosPendientes().subscribe({
+            next: (data: any) => {
+              this.ordenes = data;
+              this.items = data.map((pago: any) => ({ idOrden: pago.idOrden }));
+            },
+            error: (err) => {
+              console.error('Error al cargar órdenes de pago', err);
+            },
+          });
+          this.recargarCarrito();
+          this.id = undefined;
+        },
+        error: (err: any) => {
+          this.messagesService.errorMessage(
+            'Error',
+            'No se ha podido expirar la orden',
+          );
+        },
+      });
+      
+    }
   }
 }
